pyproject.toml
[tool.poetry]
name = "sharh"
version = "0.1.0"
description = ""
authors = ["bochikas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.4"
pillow = "10.0.0"
flake8 = "6.1.0"
flake8-noqa = "1.3.2"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.4"
psycopg2-binary = "^2.9.7"
redis = "^5.0.0"
djangorestframework-simplejwt = "^5.3.0"
django-redis = "^5.3.0"
celery = "^5.3.1"
django-filter = "^23.2"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
